Исправлен баг с размерностью номера таблицы

+++ Firmware/Source/Controller/Commutation.c
@@ -22,7 +22,7 @@ void COMM_SwitchBistableDevice(BistableSwitch Device, bool State);
 void COMM_ApplyCommutation();
 void COMM_DisableCtrlOfBistableRelay();
 void COMM_CleanShiftRegister();
-void COMM_CommutateGroupOnTableNumber(uint8_t NumbOfTable);
+void COMM_CommutateGroupOnTableNumber(uint16_t NumbOfTable);
 void COMM_CommutateForTableGroupSimpleRelay(uint64_t RelayMask);
 void COMM_CommutateForTableGroupBistablRelay(uint64_t RelayMask);
 void COMM_DisconnectAllRelay();
@@ -64,7 +64,7 @@ bool COMM_ReturnResultConnectGroup(bool *FastSwitch)
 }
 // ----------------------------------------
 
-void COMM_CommutateGroupOnTableNumber(uint8_t NumbOfTable)
+void COMM_CommutateGroupOnTableNumber(uint16_t NumbOfTable)
 {
 	COMM_CommutateForTableGroupSimpleRelay(COMM_Table[NumbOfTable].Relay & ~COMM_BUSHV);
 	COMM_CommutateForTableGroupBistablRelay(COMM_Table[NumbOfTable].Relay >> BISTABLE_RELAY_START_BIT);
